bot/requirements.txt
python-telegram-bot==22.0
httpx>=0.28.1,<1.0.0
feedparser==6.0.11
python-dotenv==1.0.1
apscheduler==3.10.4
